+++ src/carbon-calculation/clients/IEC3JobsClient.ts
@@ -3,19 +3,19 @@
 * See LICENSE.md in the project root for license terms and full copyright notice.
 *--------------------------------------------------------------------------------------------*/
 import type { AccessToken } from "@itwin/core-bentley";
-import { EC3Job, EC3JobCreate, EC3JobStatus } from "../interfaces/EC3Jobs";
+import { EC3ExtractionJobCreate, EC3Job, EC3JobStatus, EC3ReportJobCreate } from "../interfaces/EC3Jobs";
 
 export interface IEC3JobsClient {
   /**
    * Uploads report data to EC3.
    * @param {string} accessToken OAuth access token with scope `insights:modify`.
-   * @param {EC3JobCreate} job Request body.
+   * @param {EC3ReportJobCreate | EC3ExtractionJobCreate} job Request body.
    * @memberof IEC3JobsClient
    * @link https://developer.bentley.com/apis/carbon-calculation/operations/create-ec3-job/
    */
   createJob(
     accessToken: AccessToken,
-    job: EC3JobCreate
+    job: EC3ReportJobCreate | EC3ExtractionJobCreate
   ): Promise<EC3Job>;
 
   /**