+++ src/carbon-calculation/interfaces/EC3Jobs.ts
@@ -58,27 +58,36 @@ export interface EC3JobLinks {
 /**
  * Properties of the EC3Job to be created.
  * @export
- * @interface EC3JobCreate
+ * @interface EC3ReportJobCreate
  */
-export interface EC3JobCreate {
+export interface EC3ReportJobCreate {
   /**
    * EC3 Bearer token with `read write` scope. Acquiring a token is possible using EC3 APIs.
    * @type {string}
-   * @memberof EC3JobCreate
+   * @memberof EC3ReportJobCreate
    */
   ec3BearerToken: string;
   /**
    * Unique Identifier of the target EC3 Configuration.
    * @type {string}
-   * @memberof EC3JobCreate
+   * @memberof EC3ReportJobCreate
    */
   configurationId: string;
   /**
    * The name of the project to be created in EC3.
    * @type {string}
-   * @memberof EC3JobCreate
+   * @memberof EC3ReportJobCreate
    */
   projectName: string;
+}
+
+export interface EC3ExtractionJobCreate extends EC3ReportJobCreate {
+  /**
+   * Unique Identifier of the extraction to export data from.
+   * @type {string}
+   * @memberof EC3ExtractionJobCreate
+   */
+  extractionId: string;
 }
 
 /**